Revert "todo: try to fix some strange issue (?)"
This reverts commit d696972433474ff22126074745dbe42d479f9941.
