+++ src/de_mut/mod.rs
@@ -6,6 +6,7 @@ use serde::de;
 
 mod cursor;
 mod data;
+mod group;
 mod node_seq;
 mod reg;
 mod str_seq;
@@ -19,7 +20,7 @@ pub mod buildin {
 
 use cursor::{BodyCursor, Cursor, GroupCursor, PropCursor};
 use data::BorrowedValueDeserializer;
-use node_seq::NodeSeq;
+use group::GroupDeserializer;
 use r#struct::StructDeserializer;
 use reg::RegConfig;
 use structs::{RefDtb, StructureBlock, BLOCK_LEN};
@@ -146,7 +147,7 @@ impl<'de, 'b> de::MapAccess<'de> for StructAccess<'de, 'b> {
             }
             Temp::Group(cursor, len_item, len_name) => {
                 // 键是组名字，构造组反序列化器
-                seed.deserialize(&mut NodeSeq {
+                seed.deserialize(&mut GroupDeserializer {
                     dtb: self.de.dtb,
                     cursor,
                     reg: self.de.reg,
